backend/requirements.txt
# Budget App - Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
cryptography==43.0.1
aiofiles==24.1.0
